requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
authlib==1.6.9
    # via weaviate-client
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
cryptography==46.0.5
    # via authlib
deprecation==2.1.0
    # via weaviate-client
grpcio==1.78.0
    # via weaviate-client
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via weaviate-client
idna==3.11
    # via
    #   anyio
    #   httpx
packaging==26.0
    # via deprecation
protobuf==6.33.6
    # via weaviate-client
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via weaviate-client
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
validators==0.35.0
    # via weaviate-client
weaviate-client==4.20.4
    # via -r requirements.in
